# Excludes: During directory traversal, use QRegularExpression