--- a/src/libsync/discoveryphase.cpp
+++ b/src/libsync/discoveryphase.cpp
@@ -20,6 +20,7 @@
 
 #include <csync_private.h>
 #include <csync_rename.h>
+#include <csync_exclude.h>
 
 #include <QLoggingCategory>
 #include <QUrl>
@@ -717,6 +718,8 @@ void DiscoveryJob::start()
     _csync_ctx->callbacks.remote_readdir_hook = remote_vio_readdir_hook;
     _csync_ctx->callbacks.remote_closedir_hook = remote_vio_closedir_hook;
     _csync_ctx->callbacks.vio_userdata = this;
+
+    csync_exclude_traversal_prepare(_csync_ctx); // Converts the flat exclude list to optimized regexps
 
     csync_set_log_callback(_log_callback);
     csync_set_log_level(_log_level);