… test/csync/csync_tests/check_csync_exclude.cpp