@guruz 2017-09-26
Also adding the fixed patterns into a regexp is not worth it on macOS (performance regresses). https://gist.github.com/guruz/11f29b8e340d4755ccfa919d1d2bd971 
Might be worth it on Windows though with all those special Windows filenames..?