@guruz 2017-09-26
I consider this mergeable for the 2.4 alpha as soon as the last test is fixed.
```
The following tests FAILED:
	 14 - SyncFileStatusTrackerTest (OTHER_FAULT)
Errors while running CTest
```

need help for this (see IRC)