@guruz 2017-09-26
Benchmark:
```

before on master
csync_excluded_traversal: 0.005948 ms per call
csync_excluded_traversal: 0.006038 ms per call
csync_excluded_traversal: 0.005742 ms per call

after:
csync_excluded_traversal: 0.002532 ms per call
csync_excluded_traversal: 0.002361 ms per call
csync_excluded_traversal: 0.002386 ms per call
csync_excluded_traversal: 0.002445 ms per call
```