@ckamm 2017-09-27
Benchmark of f533bd95c9209dbd69d645fe3239d6982cb0bf7b on linux:

```
without patch
csync_excluded: 0.020413 ms per call
csync_excluded_traversal: 0.000756 ms per call

with patch
csync_excluded: 0.020131 ms per call
csync_excluded_traversal: 0.000650 ms per call
```

So it's not making it worse, like the old patch! :)