@ckamm 2017-09-27
I'm now much more worried about this patch: I'm not convinced our tests are sufficient to guarantee `_no_ctx` and `_traversal` both behave correctly when they use different codepaths.