@guruz 2017-09-27
superseded by https://github.com/owncloud/client/pull/6063